requirements.txt
pytest 
python-dotenv
psycopg2
requests
feedparser
beautifulsoup4
pandas
